feat: moved all types to index.d.ts
