+++ package.json
@@ -30,7 +30,7 @@
   "repository": "github:hackolade/fetch",
   "scripts": {
     "prebuild": "rimraf ./dist",
-    "build": "swc -C module.type=es6 ./src -o ./dist/esm/index.mjs && swc -C module.type=commonjs ./src -o ./dist/cjs/index.cjs",
+    "build": "swc -C module.type=es6 ./src -d ./dist/esm --out-file-extension mjs --ignore **/*.d.ts && swc -C module.type=commonjs ./src -d ./dist/cjs --out-file-extension cjs --ignore **/*.d.ts",
     "docker:compose": "docker compose --profile=test up --build --force-recreate --always-recreate-deps --remove-orphans",
     "docker:server": "npm run docker:compose -- server",
     "docker:server:verbose": "npm run docker:compose -- --attach-dependencies server",
